[Load balancers] Update nginx config to use regex
The regex match provides higher flexibility since more than one path can
be specified as part of the pattern. It also adds more specificity since
matching stops when the regex is matched, instead of the default (no)
modifier.

The location blocks need to be embedded because of the match stopping
feature.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>