[Keycloak] Use Infinispan port IPs instead of name
The Infinispan servers send their IPs as the name of the server to the
client, instead of the hostname, as defined in the inventory file. This
conflicted with the list defined for Keycloak, which caused Keycloak to
report the server's removal from the list with warning messages like:
`Server not in cluster anymore(server-name:11222), removing from the pool`.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>