[Infinispan] Add monitor user to access metrics
Configure the `monitor` user that can access the `/metrics` endpoint and
allow requests coming from the Prometheus servers to scrape the
endpoint. The requests must be authenticated using the `monitor` user
and its password.

To configure the `monitor` user, the user's password must be configured
in the vault file.

We are protecting the Infinispan metrics endpoint using an nginx reverse
proxy that only allows requests to the `/metrics` endpoint, and if they are
coming from one of the Prometheus servers. This is a complex
solution, but it is somewhat required since Infinispan appears to only
support filtering requests based on the source IP, but not for a
specific HTTP endpoint. Since the metrics endpoint is served through the
REST API that also provides access to cached data, the Prometheus servers
would also be able to attempt to query the data stored by Keycloak
(assuming they could guess the correct password) if other data endpoints
could also be accessed, consequently increasing the attack surface.

By default the metrics proxy listens on port 9000.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>