[Database] Rename systemd overrides file
The "limits.conf" name implied that the changes refer only to limits
(e.g. number of files), which was limiting in nature. The old config
file can safely be removed.

We chose to not automatically remove the old file since the new file is
later in lexicographical order, so changes to it should override the
ones in the old file.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>