[Common] Explicitly make journald store persistent
The default value on Centos for systemd-storaged's `Storage` parameter
is set to `auto`, which means that the logs are only saved to persistent
storage if the `/var/log/journal` directory is exists (it does not not
by default). Setting the value to `persistent` causes journald to create
the storage directory if it does not exist.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>