[Common] Add ZONE param in sysconfig net scripts
There is a conflict between firewalld and NetworkManager, where even if
an interface is configured as part of a zone in firewalld's
configuration files, the interface may be moved to the default zone by
NetworkManager. This change makes sure to add the ZONE field in the the
appropriate ifcfg files in `/etc/sysconfig/network-scripts`, so
NetworkManager is aware of the correct zone for the interface.

To avoid duplicating the additional rules (for both the internal and DMZ
interfaces), the rules have also been moved to a separate file that can
be included with specific parameters.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>