[Common] Update memory allocation calculations
The new way to compute memory estimates should more fairly distribute
resources when all services run on the same system. The memory
allocations for Keycloak and Infinispan are capped to 4GB to give the
database a larger chunk of memory (if available) since it stores data
for both Keycloak (i.e., configurations and event logs) and Infinispan
(i.e., persistence).

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>