[Common] Redo memory allocation calculations
Take the memory cap of 4GB for Java services into consideration when
computing the database memory allocation. This should allow for larger
memory allocations for the database on servers with more resources
(more than 16GB RAM).

NB: The Java service memory limit is only applied if services other than
Java services run on the server. Otherwise, the combined memory
allocated for Java services is equal to the system memory, minus a
reserved amount.

Also, split remaining memeory when both the database and LDAP are on the
same server. This should normally not be the case, since LDAP (if used)
should be on a separate server from Keycloak and its auxiliary services,
but consider the possiblity regardless.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>