[Rsyslog] Improve rsyslog logs, enable JSON logs
Improve rsyslog processing of logs by also reading logs from files.
This allows us to configure Keycloak and Infinispan to log in JSON
format for easier parsing, without changing the format of the logs in
the journal. Logging is configured at role level, instead of globally,
which means that only the services installed on the system have
associated rules in the rsyslog file. It is also possible to define
additional custom logging from either the journal or files.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>