[PKI] Add the "CA: true" basic constraint to certs
Configure the certificates to have the `CA: true` basic constraint and
add the "basic constraints critical" setting. This is the same behavior
as `openssl req -x509` and is a reasonable default for self-signed
certificates. The `CA: true` basic constraint is required by the
`update-ca-trust` command; otherwise, the certificate is not added to
the certificate bundle. This is especially relevant for the load
balancers' certificate, to allow clients to trust it at system level.

**NOTE**: applying this change will restart all services, as their
certificates change and the services must reload the certificates.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>