[Common] Add healthcheck to ISPN/KC services
Add health checks (check if TCP ports respond) to the Infinispan and Keycloak
systemd service files. Also remove the `PIDFile` parameter from Keycloak's
service file as the service does not create the PID file, and using it in
conjunction with the `ExecStartPost` command makes systemd believe the
service start fails.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>